added authentication & route protection to recipe book

+++ recipe-book/src/app/recipes/recipe-detail/recipe-detail.component.ts
@@ -4,6 +4,8 @@ import {ActivatedRoute, Params, Router} from '@angular/router';
 import {Recipe} from '../recipe.model';
 import {RecipeService} from '../recipe.service';
 import {Ingredient} from '../../shared/ingredient.model';
+import {AuthService} from '../../auth/auth.service';
+import {ModalService} from '../../modal.service';
 
 @Component({
   selector: 'app-recipe-detail',
@@ -13,10 +15,13 @@ import {Ingredient} from '../../shared/ingredient.model';
 export class RecipeDetailComponent implements OnInit {
   recipe: Recipe;
   id: number;
+  deleteRecipe: boolean;
 
   constructor(private recipeService: RecipeService,
               private router: Router,
-              private route: ActivatedRoute) {
+              private route: ActivatedRoute,
+              private authService: AuthService,
+              private modalService: ModalService) {
   }
 
   ngOnInit() {
@@ -39,12 +44,30 @@ export class RecipeDetailComponent implements OnInit {
   }
 
   onEditRecipe() {
-    // this.router.navigate(['edit'], {relativeTo: this.route});
+    if (!this.authService.isAuthenticated()) {
+      this.modalService.error = 'Sign up please';
+      this.modalService.show = false;
+      this.modalService.open('custom-modal-1');
+    }
     this.router.navigate(['../', this.id, 'edit'], {relativeTo: this.route});
   }
+
   onDeleteRecipe() {
-    this.recipeService.deleteRecipe(this.id);
-    this.router.navigate(['/recipes']);
+    this.modalService.show = true;
+    this.modalService.error = 'A you sure?';
+    this.modalService.open('custom-modal-1');
+    this.modalService.deleteObserver
+      .subscribe(
+        (del: boolean) => {
+          if (del) {
+            this.recipeService.deleteRecipe(this.id);
+            this.router.navigate(['/recipes']);
+            this.modalService.close('custom-modal-1');
+          } else {
+            this.modalService.close('custom-modal-1');
+          }
+        }
+      );
   }
 
 }
